Split out injector for InteractionHandler Context Menu
